module5-task1: * move generate mocks helpers
