+++ src/common/enums/index.ts
@@ -4,3 +4,4 @@ export * from './api';
 export * from './ssr';
 export * from './comment';
 export * from './app';
+export * from './mocks';
