+++ src/common/types/index.ts
@@ -1,3 +1,4 @@
 export * from './offers';
 export * from './api';
 export * from './comment';
+export * from './mocks';
