+++ src/common/types/mocks/generate-mocked-comment-cb-args.type.ts
@@ -1,0 +1,5 @@
+type GenerateMockedCommentCbArgs = {
+  comments: string[];
+};
+
+export { GenerateMockedCommentCbArgs };
