+++ src/common/types/mocks/generate-mocked-comments-cb-args.type.ts
@@ -1,0 +1,7 @@
+import { GenerateMockedCommentCbArgs } from '~/common/types/mocks/generate-mocked-comment-cb-args.type';
+
+type GenerateMockedCommentsCbArgs = {
+  count: number;
+} & GenerateMockedCommentCbArgs;
+
+export { GenerateMockedCommentsCbArgs };
