+++ src/common/types/mocks/generate-mocked-offer-cb-args.type.ts
@@ -1,0 +1,8 @@
+type GenerateMockedOfferCbArgs = {
+  titles: string[];
+  descriptions: string[];
+  categories: string[];
+  comments: string[];
+};
+
+export { GenerateMockedOfferCbArgs };
