+++ src/common/types/mocks/generate-mocked-offers-cb-args.type.ts
@@ -1,0 +1,7 @@
+import { GenerateMockedOfferCbArgs } from '~/common/types/mocks/generate-mocked-offer-cb-args.type';
+
+type GenerateMockedOffersCbArgs = {
+  count: number;
+} & GenerateMockedOfferCbArgs;
+
+export { GenerateMockedOffersCbArgs };
