+++ src/common/types/mocks/index.ts
@@ -1,0 +1,4 @@
+export * from './generate-mocked-comment-cb-args.type';
+export * from './generate-mocked-comments-cb-args.type';
+export * from './generate-mocked-offer-cb-args.type';
+export * from './generate-mocked-offers-cb-args.type';
