+++ src/helpers/index.ts
@@ -5,3 +5,4 @@ export * from './fs';
 export * from './validation';
 export * from './log';
 export * from './file';
+export * from './mocks';
