+++ src/helpers/mocks/get-mocked-offers-data.helper.ts
@@ -1,0 +1,25 @@
+import { MocksConfig } from '~/common/enums';
+import { GenerateMockedOfferCbArgs } from '~/common/types';
+import { readOfferFileContent } from '~/helpers/mocks/read-offer-file-content.helper';
+
+const dataPaths = [
+  MocksConfig.TITLE.FILE_PATH,
+  MocksConfig.DESCRIPTION.FILE_PATH,
+  MocksConfig.CATEGORY.FILE_PATH,
+  MocksConfig.COMMENTS.FILE_PATH,
+];
+
+const getMockedOffersData = async (): Promise<GenerateMockedOfferCbArgs> => {
+  const [titles, descriptions, categories, comments] = await Promise.all(
+    dataPaths.map((path) => readOfferFileContent(path)),
+  );
+
+  return {
+    titles,
+    descriptions,
+    categories,
+    comments,
+  };
+};
+
+export { getMockedOffersData };
