+++ src/service/cli/generate/helpers.ts
@@ -1,147 +1,22 @@
-import {
-  paintMessage,
-  writeToFile,
-  readFile,
-  getRandomItem,
-  getTwoDigitalString,
-  getRandomNumber,
-  getRandomItems,
-  getRandomId,
-  logger,
-} from '~/helpers';
-import { OfferType } from '~/common/enums';
-import { IOffer, IComment } from '~/common/interfaces';
-import {
-  GenerateMockedCommentCbArgs,
-  GenerateMockedCommentsCbArgs,
-  GenerateMockedOfferCbArgs,
-  GenerateMockedOffersCbArgs,
-  MocksConfig,
-} from './common';
-
-const offerTypes = Object.values(OfferType);
-
-const generateMockedComment = ({
-  comments,
-}: GenerateMockedCommentCbArgs): IComment => ({
-  id: getRandomId(),
-  text: getRandomItems(
-    comments,
-    getRandomNumber(
-      MocksConfig.COMMENTS.MIN_SENTENCES_COUNT,
-      MocksConfig.COMMENTS.MAX_SENTENCES_COUNT,
-    ),
-  ).join(` `),
-});
-
-const generateMockedComments = ({
-  count,
-  comments,
-}: GenerateMockedCommentsCbArgs): IComment[] => {
-  const mockedComments = Array.from(new Array(count), () => (
-    generateMockedComment({ comments })
-  ));
-
-  return mockedComments;
-};
-
-const generateMockedOffer = ({
-  titles,
-  descriptions,
-  categories,
-  comments,
-}: GenerateMockedOfferCbArgs): IOffer => ({
-  id: getRandomId(),
-  title: getRandomItem(titles),
-  picture: `item${getTwoDigitalString(
-    getRandomNumber(
-      MocksConfig.PICTURE_NUMBER.MIN,
-      MocksConfig.PICTURE_NUMBER.MAX,
-    ),
-  )}.jpg`,
-  description: getRandomItems(
-    descriptions,
-    getRandomNumber(
-      MocksConfig.DESCRIPTION.MIN_COUNT,
-      MocksConfig.DESCRIPTION.MAX_COUNT,
-    ),
-  ).join(` `),
-  type: getRandomItem(offerTypes),
-  sum: getRandomNumber(MocksConfig.PRICE.MIN, MocksConfig.PRICE.MAX),
-  category: getRandomItems(
-    categories,
-    getRandomNumber(MocksConfig.CATEGORY.MIN_COUNT, categories.length),
-  ),
-  comments: generateMockedComments({
-    count: getRandomNumber(
-      MocksConfig.COMMENTS.MIN_COUNT,
-      MocksConfig.COMMENTS.MAX_COUNT,
-    ),
-    comments,
-  }),
-});
-
-const generateMockedOffers = ({
-  count,
-  titles,
-  categories,
-  descriptions,
-  comments,
-}: GenerateMockedOffersCbArgs): IOffer[] => {
-  const mockedOffers = Array.from(new Array(count), () => (
-    generateMockedOffer({ titles, categories, descriptions, comments })
-  ));
-
-  return mockedOffers;
-};
+import { paintMessage, writeToFile, logger } from '~/helpers';
+import { MocksConfig } from '~/common/enums';
+import { IOffer } from '~/common/interfaces';
 
 const saveOffersToFile = async (mockedOffers: IOffer[]): Promise<void> => {
   try {
     await writeToFile(MocksConfig.FILE_NAME, JSON.stringify(mockedOffers));
 
-    logger.info(paintMessage(`Operation success. File with mocks was created.`, `green`));
-  } catch {
-    logger.error(paintMessage(`An error occurred on saving mocked-data: can't write mocked-data to file...`, `red`));
-  }
-};
-
-const readOfferFileContent = async (path: string): Promise<string[]> => {
-  try {
-    const content = await readFile(path);
-
-    return content.trim().split(`\n`);
+    logger.info(
+      paintMessage(`Operation success. File with mocks was created.`, `green`),
+    );
   } catch {
     logger.error(
       paintMessage(
-        `An error occurred on reading mocked-data: can't read mocked-data from file...`,
+        `An error occurred on saving mocked-data: can't write mocked-data to file...`,
         `red`,
       ),
     );
-
-    return [];
   }
 };
 
-const getOffersData = async (): Promise<GenerateMockedOfferCbArgs> => {
-  const titles = await readOfferFileContent(MocksConfig.TITLE.FILE_PATH);
-  const descriptions = await readOfferFileContent(
-    MocksConfig.DESCRIPTION.FILE_PATH,
-  );
-  const categories = await readOfferFileContent(MocksConfig.CATEGORY.FILE_PATH);
-  const comments = await readOfferFileContent(MocksConfig.COMMENTS.FILE_PATH);
-
-  return {
-    titles,
-    descriptions,
-    categories,
-    comments,
-  };
-};
-
-export {
-  generateMockedOffer,
-  generateMockedOffers,
-  saveOffersToFile,
-  readOfferFileContent,
-  getOffersData,
-};
+export { saveOffersToFile };
